Moving VitisBuild and SlashBuild into a common alveo_build module
