+++ src/finn/builder/build_dataflow_config.py
@@ -34,7 +34,7 @@ from dataclasses_json import dataclass_json
 from enum import Enum
 from typing import Any, List, Optional
 
-from finn.transformation.fpgadataflow.vitis_build import VitisOptStrategy
+from finn.transformation.fpgadataflow.alveo_build import VitisOptStrategy
 from finn.util.basic import part_map, vitis_default_platform
