+++ tests/fpgadataflow/test_fpgadataflow_ipstitch.py
@@ -51,7 +51,7 @@ from finn.transformation.fpgadataflow.insert_tlastmarker import InsertTLastMarke
 from finn.transformation.fpgadataflow.make_zynq_proj import ZynqBuild
 from finn.transformation.fpgadataflow.prepare_ip import PrepareIP
 from finn.transformation.fpgadataflow.synth_ooc import SynthOutOfContext
-from finn.transformation.fpgadataflow.vitis_build import VitisBuild
+from finn.transformation.fpgadataflow.alveo_build import VitisBuild
 from finn.util.basic import pynq_part_map, vitis_default_platform, vitis_part_map
 from finn.util.test import load_test_checkpoint_or_skip
